requirements.txt
mysql-connector-python==9.0.0
Pillow==11.0.0

